added local markets and international markets

+++ database/migrations/2019_08_07_082220_create_loosers_table.php
@@ -17,7 +17,7 @@ class CreateLoosersTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->integer('price');
-            $table->integer('percentage');
+            $table->decimal('percentage');
             $table->timestamps();
         });
     }
